Move hir::Lit -> ty::Const conversion into its own file
